requirements.txt
asgiref==3.6.0
autopep8==2.0.1
crispy-bootstrap5==0.7
Django==4.1.7
django-ckeditor==6.5.1
django-crispy-forms==2.0
django-js-asset==2.0.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
phonenumbers==8.13.6
Pillow==9.4.0
pycodestyle==2.10.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
